+++ ops/op2/test_cases/async/async_arg_return_result.out
@@ -47,7 +47,7 @@ pub const fn op_async() -> ::deno_core::_ops::OpDecl {
             let result = {
                 let arg0 = args.get(1usize as i32);
                 let Some(arg0) = deno_core::_ops::to_i32_option(&arg0) else {
-                    deno_core::_ops::throw_error1(&info, "expected i32");
+                    deno_core::_ops::throw_error_one_byte_info(&info, "expected i32");
                     return 1;
                 };
                 let arg0 = arg0 as _;
