+++ ops/op2/test_cases/async/async_jsbuffer.out
@@ -51,7 +51,7 @@ pub const fn op_async_v8_buffer() -> ::deno_core::_ops::OpDecl {
                 arg0_temp = match unsafe { deno_core::_ops::to_v8_slice::<u8>(arg0) } {
                     Ok(arg0) => arg0,
                     Err(arg0_err) => {
-                        deno_core::_ops::throw_error1(&info, arg0_err);
+                        deno_core::_ops::throw_error_one_byte_info(&info, arg0_err);
                         return 1;
                     }
                 };
@@ -76,7 +76,7 @@ pub const fn op_async_v8_buffer() -> ::deno_core::_ops::OpDecl {
                 ) {
                     Ok(v) => rv.set(v),
                     Err(rv_err) => {
-                        deno_core::_ops::throw_error2(&mut scope, rv_err);
+                        deno_core::_ops::throw_error_anyhow(&mut scope, rv_err);
                         return 1;
                     }
                 };
