+++ ops/op2/test_cases/sync/add.out
@@ -142,13 +142,13 @@ const fn op_add() -> ::deno_core::_ops::OpDecl {
             let result = {
                 let arg0 = args.get(0usize as i32);
                 let Some(arg0) = deno_core::_ops::to_u32_option(&arg0) else {
-                    deno_core::_ops::throw_error1(&info, "expected u32");
+                    deno_core::_ops::throw_error_one_byte_info(&info, "expected u32");
                     return 1;
                 };
                 let arg0 = arg0 as _;
                 let arg1 = args.get(1usize as i32);
                 let Some(arg1) = deno_core::_ops::to_u32_option(&arg1) else {
-                    deno_core::_ops::throw_error1(&info, "expected u32");
+                    deno_core::_ops::throw_error_one_byte_info(&info, "expected u32");
                     return 1;
                 };
                 let arg1 = arg1 as _;
