+++ ops/op2/test_cases/sync/add_options.out
@@ -41,7 +41,7 @@ pub const fn op_test_add_option() -> ::deno_core::_ops::OpDecl {
             let result = {
                 let arg0 = args.get(0usize as i32);
                 let Some(arg0) = deno_core::_ops::to_u32_option(&arg0) else {
-                    deno_core::_ops::throw_error1(&info, "expected u32");
+                    deno_core::_ops::throw_error_one_byte_info(&info, "expected u32");
                     return 1;
                 };
                 let arg0 = arg0 as _;
@@ -50,7 +50,10 @@ pub const fn op_test_add_option() -> ::deno_core::_ops::OpDecl {
                     None
                 } else {
                     let Some(arg1) = deno_core::_ops::to_u32_option(&arg1) else {
-                        deno_core::_ops::throw_error1(&info, "expected u32");
+                        deno_core::_ops::throw_error_one_byte_info(
+                            &info,
+                            "expected u32",
+                        );
                         return 1;
                     };
                     let arg1 = arg1 as _;
