+++ ops/op2/test_cases/sync/buffers.out
@@ -151,7 +151,7 @@ const fn op_buffers() -> ::deno_core::_ops::OpDecl {
                 arg0_temp = match unsafe { deno_core::_ops::to_v8_slice::<u8>(arg0) } {
                     Ok(arg0) => arg0,
                     Err(arg0_err) => {
-                        deno_core::_ops::throw_error1(&info, arg0_err);
+                        deno_core::_ops::throw_error_one_byte_info(&info, arg0_err);
                         return 1;
                     }
                 };
@@ -161,7 +161,7 @@ const fn op_buffers() -> ::deno_core::_ops::OpDecl {
                 arg1_temp = match unsafe { deno_core::_ops::to_v8_slice::<u8>(arg1) } {
                     Ok(arg1) => arg1,
                     Err(arg1_err) => {
-                        deno_core::_ops::throw_error1(&info, arg1_err);
+                        deno_core::_ops::throw_error_one_byte_info(&info, arg1_err);
                         return 1;
                     }
                 };
@@ -171,7 +171,7 @@ const fn op_buffers() -> ::deno_core::_ops::OpDecl {
                 arg2_temp = match unsafe { deno_core::_ops::to_v8_slice::<u8>(arg2) } {
                     Ok(arg2) => arg2,
                     Err(arg2_err) => {
-                        deno_core::_ops::throw_error1(&info, arg2_err);
+                        deno_core::_ops::throw_error_one_byte_info(&info, arg2_err);
                         return 1;
                     }
                 };
@@ -185,7 +185,7 @@ const fn op_buffers() -> ::deno_core::_ops::OpDecl {
                 arg3_temp = match unsafe { deno_core::_ops::to_v8_slice::<u8>(arg3) } {
                     Ok(arg3) => arg3,
                     Err(arg3_err) => {
-                        deno_core::_ops::throw_error1(&info, arg3_err);
+                        deno_core::_ops::throw_error_one_byte_info(&info, arg3_err);
                         return 1;
                     }
                 };
@@ -394,7 +394,7 @@ const fn op_buffers_32() -> ::deno_core::_ops::OpDecl {
                 arg0_temp = match unsafe { deno_core::_ops::to_v8_slice::<u32>(arg0) } {
                     Ok(arg0) => arg0,
                     Err(arg0_err) => {
-                        deno_core::_ops::throw_error1(&info, arg0_err);
+                        deno_core::_ops::throw_error_one_byte_info(&info, arg0_err);
                         return 1;
                     }
                 };
@@ -404,7 +404,7 @@ const fn op_buffers_32() -> ::deno_core::_ops::OpDecl {
                 arg1_temp = match unsafe { deno_core::_ops::to_v8_slice::<u32>(arg1) } {
                     Ok(arg1) => arg1,
                     Err(arg1_err) => {
-                        deno_core::_ops::throw_error1(&info, arg1_err);
+                        deno_core::_ops::throw_error_one_byte_info(&info, arg1_err);
                         return 1;
                     }
                 };
@@ -414,7 +414,7 @@ const fn op_buffers_32() -> ::deno_core::_ops::OpDecl {
                 arg2_temp = match unsafe { deno_core::_ops::to_v8_slice::<u32>(arg2) } {
                     Ok(arg2) => arg2,
                     Err(arg2_err) => {
-                        deno_core::_ops::throw_error1(&info, arg2_err);
+                        deno_core::_ops::throw_error_one_byte_info(&info, arg2_err);
                         return 1;
                     }
                 };
@@ -428,7 +428,7 @@ const fn op_buffers_32() -> ::deno_core::_ops::OpDecl {
                 arg3_temp = match unsafe { deno_core::_ops::to_v8_slice::<u32>(arg3) } {
                     Ok(arg3) => arg3,
                     Err(arg3_err) => {
-                        deno_core::_ops::throw_error1(&info, arg3_err);
+                        deno_core::_ops::throw_error_one_byte_info(&info, arg3_err);
                         return 1;
                     }
                 };
@@ -535,7 +535,7 @@ const fn op_buffers_option() -> ::deno_core::_ops::OpDecl {
                     } {
                         Ok(arg0) => arg0,
                         Err(arg0_err) => {
-                            deno_core::_ops::throw_error1(&info, arg0_err);
+                            deno_core::_ops::throw_error_one_byte_info(&info, arg0_err);
                             return 1;
                         }
                     };
@@ -552,7 +552,7 @@ const fn op_buffers_option() -> ::deno_core::_ops::OpDecl {
                     } {
                         Ok(arg1) => arg1,
                         Err(arg1_err) => {
-                            deno_core::_ops::throw_error1(&info, arg1_err);
+                            deno_core::_ops::throw_error_one_byte_info(&info, arg1_err);
                             return 1;
                         }
                     };
