+++ ops/op2/test_cases/sync/buffers_out.out
@@ -45,7 +45,7 @@ const fn op_buffers() -> ::deno_core::_ops::OpDecl {
                 arg0_temp = match unsafe { deno_core::_ops::to_v8_slice::<u8>(arg0) } {
                     Ok(arg0) => arg0,
                     Err(arg0_err) => {
-                        deno_core::_ops::throw_error1(&info, arg0_err);
+                        deno_core::_ops::throw_error_one_byte_info(&info, arg0_err);
                         return 1;
                     }
                 };
@@ -55,7 +55,7 @@ const fn op_buffers() -> ::deno_core::_ops::OpDecl {
             match deno_core::_ops::RustToV8Fallible::to_v8_fallible(result, &mut scope) {
                 Ok(v) => rv.set(v),
                 Err(rv_err) => {
-                    deno_core::_ops::throw_error2(&mut scope, rv_err);
+                    deno_core::_ops::throw_error_anyhow(&mut scope, rv_err);
                     return 1;
                 }
             };
@@ -157,7 +157,7 @@ const fn op_buffers_option() -> ::deno_core::_ops::OpDecl {
                     } {
                         Ok(arg0) => arg0,
                         Err(arg0_err) => {
-                            deno_core::_ops::throw_error1(&info, arg0_err);
+                            deno_core::_ops::throw_error_one_byte_info(&info, arg0_err);
                             return 1;
                         }
                     };
@@ -169,7 +169,7 @@ const fn op_buffers_option() -> ::deno_core::_ops::OpDecl {
             match deno_core::_ops::RustToV8Fallible::to_v8_fallible(result, &mut scope) {
                 Ok(v) => rv.set(v),
                 Err(rv_err) => {
-                    deno_core::_ops::throw_error2(&mut scope, rv_err);
+                    deno_core::_ops::throw_error_anyhow(&mut scope, rv_err);
                     return 1;
                 }
             };
@@ -266,7 +266,7 @@ const fn op_arraybuffers() -> ::deno_core::_ops::OpDecl {
                 arg0_temp = match unsafe { deno_core::_ops::to_v8_slice_buffer(arg0) } {
                     Ok(arg0) => arg0,
                     Err(arg0_err) => {
-                        deno_core::_ops::throw_error1(&info, arg0_err);
+                        deno_core::_ops::throw_error_one_byte_info(&info, arg0_err);
                         return 1;
                     }
                 };
