+++ ops/op2/test_cases/sync/from_v8.out
@@ -44,7 +44,7 @@ pub const fn op_from_v8_arg() -> ::deno_core::_ops::OpDecl {
                 let arg0 = match <Foo as deno_core::FromV8>::from_v8(&mut scope, arg0) {
                     Ok(t) => t,
                     Err(arg0_err) => {
-                        deno_core::_ops::throw_error2(&mut scope, arg0_err);
+                        deno_core::_ops::throw_error_anyhow(&mut scope, arg0_err);
                         return 1;
                     }
                 };
