+++ ops/op2/test_cases/sync/object_wrap.out
@@ -63,7 +63,10 @@ impl Foo {
                         None
                     } else {
                         let Some(arg0) = deno_core::_ops::to_u32_option(&arg0) else {
-                            deno_core::_ops::throw_error1(&info, "expected u32");
+                            deno_core::_ops::throw_error_one_byte_info(
+                                &info,
+                                "expected u32",
+                            );
                             return 1;
                         };
                         let arg0 = arg0 as _;
@@ -178,7 +181,7 @@ impl Foo {
                 let Some(self_) = deno_core::_ops::try_unwrap_cppgc_object::<
                     Foo,
                 >(&mut scope, args.this().into()) else {
-                    deno_core::_ops::throw_error1(&info, "expected Foo");
+                    deno_core::_ops::throw_error_one_byte_info(&info, "expected Foo");
                     return 1;
                 };
                 let self_ = &*self_;
@@ -277,14 +280,17 @@ impl Foo {
                 let Some(self_) = deno_core::_ops::try_unwrap_cppgc_object::<
                     Foo,
                 >(&mut scope, args.this().into()) else {
-                    deno_core::_ops::throw_error1(&info, "expected Foo");
+                    deno_core::_ops::throw_error_one_byte_info(&info, "expected Foo");
                     return 1;
                 };
                 let self_ = &*self_;
                 let result = {
                     let arg0 = args.get(0usize as i32);
                     let Some(arg0) = deno_core::_ops::to_u32_option(&arg0) else {
-                        deno_core::_ops::throw_error1(&info, "expected u32");
+                        deno_core::_ops::throw_error_one_byte_info(
+                            &info,
+                            "expected u32",
+                        );
                         return 1;
                     };
                     let arg0 = arg0 as _;
@@ -402,14 +408,17 @@ impl Foo {
                 let Some(self_) = deno_core::_ops::try_unwrap_cppgc_object::<
                     Foo,
                 >(&mut scope, args.this().into()) else {
-                    deno_core::_ops::throw_error1(&info, "expected Foo");
+                    deno_core::_ops::throw_error_one_byte_info(&info, "expected Foo");
                     return 1;
                 };
                 let self_ = &*self_;
                 let result = {
                     let arg0 = args.get(0usize as i32);
                     let Some(arg0) = deno_core::_ops::to_u32_option(&arg0) else {
-                        deno_core::_ops::throw_error1(&info, "expected u32");
+                        deno_core::_ops::throw_error_one_byte_info(
+                            &info,
+                            "expected u32",
+                        );
                         return 1;
                     };
                     let arg0 = arg0 as _;
@@ -566,7 +575,10 @@ impl Foo {
                                 &*fast_api_callback_options,
                             )
                         };
-                        deno_core::_ops::throw_error3(&mut scope, "expected Foo");
+                        deno_core::_ops::throw_error_one_byte(
+                            &mut scope,
+                            "expected Foo",
+                        );
                         return unsafe { std::mem::zeroed() };
                     }
                 };
@@ -591,7 +603,7 @@ impl Foo {
                 let Some(self_) = deno_core::_ops::try_unwrap_cppgc_object::<
                     Foo,
                 >(&mut scope, args.this().into()) else {
-                    deno_core::_ops::throw_error1(&info, "expected Foo");
+                    deno_core::_ops::throw_error_one_byte_info(&info, "expected Foo");
                     return 1;
                 };
                 let self_ = &*self_;
@@ -691,7 +703,7 @@ impl Foo {
                 let Some(self_) = deno_core::_ops::try_unwrap_cppgc_object::<
                     Foo,
                 >(&mut scope, args.this().into()) else {
-                    deno_core::_ops::throw_error1(&info, "expected Foo");
+                    deno_core::_ops::throw_error_one_byte_info(&info, "expected Foo");
                     return 1;
                 };
                 let self_ = &*self_;
@@ -794,7 +806,7 @@ impl Foo {
                 let Some(self_) = deno_core::_ops::try_unwrap_cppgc_object::<
                     Foo,
                 >(&mut scope, args.this().into()) else {
-                    deno_core::_ops::throw_error1(&info, "expected Foo");
+                    deno_core::_ops::throw_error_one_byte_info(&info, "expected Foo");
                     return 1;
                 };
                 let self_ = &*self_;
@@ -980,7 +992,7 @@ impl Foo {
                 let Some(self_) = deno_core::_ops::try_unwrap_cppgc_object::<
                     Foo,
                 >(&mut scope, args.this().into()) else {
-                    deno_core::_ops::throw_error1(&info, "expected Foo");
+                    deno_core::_ops::throw_error_one_byte_info(&info, "expected Foo");
                     return 1;
                 };
                 let self_ = &*self_;
