+++ ops/op2/test_cases/sync/string_onebyte.out
@@ -127,7 +127,7 @@ const fn op_string_onebyte() -> ::deno_core::_ops::OpDecl {
                 let arg0 = match deno_core::_ops::to_cow_one_byte(&mut scope, &arg0) {
                     Ok(arg0) => arg0,
                     Err(arg0) => {
-                        deno_core::_ops::throw_error1(&info, arg0);
+                        deno_core::_ops::throw_error_one_byte_info(&info, arg0);
                         return 1;
                     }
                 };
