+++ ops/op2/test_cases/sync/string_option_return.out
@@ -51,7 +51,7 @@ pub const fn op_string_return() -> ::deno_core::_ops::OpDecl {
             match deno_core::_ops::RustToV8Fallible::to_v8_fallible(result, &mut scope) {
                 Ok(v) => rv.set(v),
                 Err(rv_err) => {
-                    deno_core::_ops::throw_error2(&mut scope, rv_err);
+                    deno_core::_ops::throw_error_anyhow(&mut scope, rv_err);
                     return 1;
                 }
             };
