+++ ops/op2/test_cases/sync/v8_global.out
@@ -44,7 +44,7 @@ pub const fn op_global() -> ::deno_core::_ops::OpDecl {
                 let Ok(mut arg0) = deno_core::_ops::v8_try_convert::<
                     deno_core::v8::String,
                 >(arg0) else {
-                    deno_core::_ops::throw_error1(&info, "expected String");
+                    deno_core::_ops::throw_error_one_byte_info(&info, "expected String");
                     return 1;
                 };
                 let arg0 = deno_core::v8::Global::new(&mut scope, arg0);
