+++ ops/op2/test_cases/sync/v8_ref_option.out
@@ -114,7 +114,10 @@ pub const fn op_v8_lifetime() -> ::deno_core::_ops::OpDecl {
                                 &*fast_api_callback_options,
                             )
                         };
-                        deno_core::_ops::throw_error3(&mut scope, "expected String");
+                        deno_core::_ops::throw_error_one_byte(
+                            &mut scope,
+                            "expected String",
+                        );
                         return unsafe { std::mem::zeroed() };
                     }
                 };
@@ -131,7 +134,10 @@ pub const fn op_v8_lifetime() -> ::deno_core::_ops::OpDecl {
                                 &*fast_api_callback_options,
                             )
                         };
-                        deno_core::_ops::throw_error3(&mut scope, "expected String");
+                        deno_core::_ops::throw_error_one_byte(
+                            &mut scope,
+                            "expected String",
+                        );
                         return unsafe { std::mem::zeroed() };
                     }
                 };
@@ -159,7 +165,7 @@ pub const fn op_v8_lifetime() -> ::deno_core::_ops::OpDecl {
                 let Ok(mut arg0) = deno_core::_ops::v8_try_convert_option::<
                     deno_core::v8::String,
                 >(arg0) else {
-                    deno_core::_ops::throw_error1(&info, "expected String");
+                    deno_core::_ops::throw_error_one_byte_info(&info, "expected String");
                     return 1;
                 };
                 let arg0 = match &arg0 {
@@ -170,7 +176,7 @@ pub const fn op_v8_lifetime() -> ::deno_core::_ops::OpDecl {
                 let Ok(mut arg1) = deno_core::_ops::v8_try_convert_option::<
                     deno_core::v8::String,
                 >(arg1) else {
-                    deno_core::_ops::throw_error1(&info, "expected String");
+                    deno_core::_ops::throw_error_one_byte_info(&info, "expected String");
                     return 1;
                 };
                 let arg1 = match &arg1 {
