+++ ops/op2/test_cases/sync/v8_string.out
@@ -43,7 +43,7 @@ const fn op_v8_string() -> ::deno_core::_ops::OpDecl {
                 let Ok(mut arg0) = deno_core::_ops::v8_try_convert::<
                     deno_core::v8::String,
                 >(arg0) else {
-                    deno_core::_ops::throw_error1(&info, "expected String");
+                    deno_core::_ops::throw_error_one_byte_info(&info, "expected String");
                     return 1;
                 };
                 let arg0 = &arg0;
@@ -51,7 +51,7 @@ const fn op_v8_string() -> ::deno_core::_ops::OpDecl {
                 let Ok(mut arg1) = deno_core::_ops::v8_try_convert::<
                     deno_core::v8::String,
                 >(arg1) else {
-                    deno_core::_ops::throw_error1(&info, "expected String");
+                    deno_core::_ops::throw_error_one_byte_info(&info, "expected String");
                     return 1;
                 };
                 let arg1 = arg1;
